Server/requirements.txt
Flask
PyJWT
pyotp
python-dotenv
qrcode
Requests
urllib3
Pillow
